Include IDF in new document postings
Forgot to make an update to the document and database layer to calculate the inverse document frequency upon insertion of new documents.  If you've been marking sites and wondering why search doesn't work, this is why.  Resubmit the sites, and they should work in search again.  Sorry for the inconvenience.